git-svn-id: https://subversion.assembla.com/svn/smart-cs/trunk@5428 37136b1e-4cb1-4992-8394-573b48a9bdce

+++ source/extensions/i2/org.wcs.smart.i2/src/org/wcs/smart/i2/ui/editors/record/RecordSummaryPage.java
@@ -421,6 +421,7 @@ public class RecordSummaryPage extends EditorPart{
 	}
 	
 	public void setEditMode(boolean editMode){		
+		if (editItem.isDisposed()) return;
 		editItem.setSelection(editMode);
 		if (IntelSecurityManager.INSTANCE.canDeleteRecord()){
 			deleteItem.setEnabled(editMode);
